Disables submit button in SelectionPanel To prevent the form from being submit without a location, the submit button is disabled until the `coordinates` prop has a value.
Due to the nature of the form validation, the 'Het object staat niet op de kaart' checkbox can be checked and the form input is then considered valid. With `react-reactive-form` it is not possible to define a complex, compound validation that evaluates the field's value as well as external values. In this case, the location (coordinates and address) and the selection on the map. To make sure that the interaction with the map will always result in having coordinates set, there is no other way than to disable the submit button.