(Fix) iOS AddressPanel scroll This PR fixes a scroll issue on iOS where, when an input element inside a fixed position element gets focus, the OS keyboard will push the fixed positioned element off-screen. To solve this, there are two options: 1. Turn all fixed elements into absolute positioned elements 2. Do not let the input field get focus
The first option is quite troublesome, because it requires a refactoring of most of the styled component that make up the detail panel and the address panel. The second option is the easiest, but it requires the user to tap the input field twice; once to toggle the full-screen panel and the second time to focus the input field.

For now, the second option is chosen. It this is considered to be too obtrusive, the first option will be implemented.