[SIG-4810] Subcategorie met objecten tonen in het filterpanel (#2333)
Context:
Filtering on subCategories in Filterpanel needs to be possible. It is not allowed to have selected a subCategory and not
also its mainCategory. Icons of subCategory, if they are available, needs to be shown. Otherwise that of the mainCategory
and if that's not available the standard marker needs to be shown. Sizes of the markers and icons will be addressed in a
seperate story. It is allowed to have checked a mainCategory and not one of its subCategories. 
Only of the mainCategories Afval and Wegen... the subCategories are shown.
Only filters of which there are actual incidents in the map are shown. 

Changes:
- added a FilterCategories where it is decided whether there are subCategories and if the mainCategories are Afval and
Wegen... of if only the mainCategory needs to be shown.
- added a FilterCategory which contains the checkbox, icon and name of category
- added functionality to exclude subcategories when maincategory is not checked
- added functionality to compute if there are incidents in map and if so, to adjust filters accordingly.

* SIG-4664_create_filter_in_FilterPanel_in_Meldingenkaart