[fix-4981]: removed Dashboardlink from login page
Add isAuthenticated check for showing Dashboard.

In the backoffice, where one can enter credentials, the menu item dashboard was visible. This should not be visible when a user is not logged in. Now it is only visisble if user is authenticated. Test need not be added, because there is already a test for renderingen when not authenticated.