Add general Alert on first page where a municipality can set the text
Municipalities want to have to opportunity to place an alert on the first page of the app. They also want to be able to set it itself and publish it itself without needing Amsterdam. So the alert is done as a object in the config files app.base.json and app.<municipality name>.json. Municipalities can than insert text into their config file without needing us.

Configuration:
-  added to app.base.json an empty frontPageAlert
-  added to app.amsterdam.json a frontPageAlert which is not empty so that it can be tested.

Changes:
-  In config files added frontPageAlert
-  Added a FrontPageAlert which only shows if there is text in the config file at frontPageAlert
-  Show in IncidentWizard when on first page the FrontPageAlert