[feat-5050]: Add general Alert on first page (#2521)
Context
Municipalities want to have to opportunity to place an alert on the first page of the app. They also want to be able to set it itself and publish it itself without needing Amsterdam. So the alert is done as a object in the config files app.base.json and app.<municipality name>.json. Municipalities can than insert text into their config file without needing us.

Configuration:
-  added to app.base.json an empty frontPageAlert


Changes:
-  In config files added frontPageAlert
-  Added a FrontPageAlert which only shows if there is text in the config file at frontPageAlert
-  Show in IncidentWizard when on first page the FrontPageAler