import fonts.css in app.js
Fonts removed from asset folder in order to prevent doubles on the server.
Fonts.css needed to be included in app.js for webpack to know it's there.
Fonts.css has been moved next to global.css in order to prevent an error in the path (with 4 dots in path)