make code prettier in statusForm
Context
If  everything is done in Statusform and as much as possible in one function, then code will be more maintainable in the long run. So code has been reworked to check for the change and cause a globalNotification in the onStatusChange in the statusform. The Verstuur button has been disabled if category is Overig - overig and people change the status from another status to afgehandeld.