[feat-5083] stop processing when category is Overig-overig and status is Afgehandeld (#2544)
**Context**
An issue cannot get the status Afgehandeld when it's category is Overig-overig. The back-end still needs to be able to accept this, but the back-office should prevent this.

**Changes**
In Statusform in the OnStatusChange there is a check added for the categories and the status. Also the button Verstuur is disabled when the same checks are passed. 

Co-authored-by: Oscar <bakkeroscar@gmail.com>