feat/5450-remove-inconsistencies-in-styling-of-forms (#2711)
Made the space between geschiedenis and the inputfield for Notitie 32 px. 
Made sure that in incidentdetail of Incident-Management the spacing remained the same. 
Made the spacing between Standaardteksten per status and the select 8px. 
Checked the spacing of the other components. They are either 4 px or compliant with the spacing used in their component.