[feat-5240]: 'instellingen' is included in url of nested pages (#2723)
* instellingen included in url

Because wildcards are now possible in Route, the strings in route are no longer prefixed with an /. They will append to url instead of replace.
Because the strings in route no longer have an /, the navigate_to need to have the full url to navigate to. So all, including tests, are prefixed with the base_url of settings.
Redirected all wrong url's to the page not found page.