Added useWatch to handle dependency of component on another component… (#2734)
… in StandardTextFiled

Because the parentcomponent Detail is not always rendered, the formValues are not always updated. React Forms has useWatch to check for changing of values in the form. FormValues are therefor removed and useWatch is implemented in CategoryForm where the showing of the standardTextFields is dependent on the status and the chosen standard text. test coverage of StandardTextsFiled improved. Introduced watch on is_public_name as well.tegoryForm where the showing of the standardTextFields is dependent on the status and the chosen standard text.

Fixed ES Lint Warning in constants