Revert "Added useWatch to handle dependency of component on another component… (#2734)" (#2738)
This reverts commit 66ae7c84645ed893533986062ebffe796343fa8a.
Reverting because the bug it was supposed to fix (filtering Standard Texts on its status Active/Inactive) isn't what that status is. The status in Instellingen -> Subcategories -> Subcategory is the status of the subcategory (and not of the Standardtexts associated with that subcategory). IN story SIG-5536 in Jira the desired functionality is requested.